Revert "git-clone.txt: remove the restriction on pushing from a shallow clone"
This reverts commit cbf02fca11d11a49f3538a981badfb8e37011849.

"It fails reliably without corrupting the receiving repository when
it should fail" may be better than the situation before the receiving
end was hardened recently, but the fact that sometimes the push does
not go through still remains.  It is better to advice the users that
they cannot push from a shallow repository as a limitation before
they decide to use (or not to use) a shallow clone.

Signed-off-by: Junio C Hamano <gitster@pobox.com>