Revert "Add new @ shortcut for HEAD"
This reverts commit 8e0f3a655c8ec8ce322eb9c24bee7fcc257eadae, as it
does not just apply to "@" (and forms with modifiers like @{u}
applied to it), but also affects e.g. "refs/heads/@/foo", which it
shouldn't.

The basic idea of giving a short-hand might be good, and the topic
can be retried later, but let's revert to avoid affecting existing
use cases for now for the upcoming release.